add maxAliasForCollections override via Parser case class
